+++ package.json
@@ -41,7 +41,7 @@
     "@radix-ui/react-switch": "^1.0.3",
     "@radix-ui/react-tabs": "^1.1.0",
     "@radix-ui/react-toast": "^1.2.1",
-    "@radix-ui/react-tooltip": "^1.0.7",
+    "@radix-ui/react-tooltip": "^1.1.2",
     "@tanstack/react-table": "^8.13.2",
     "add": "^2.0.6",
     "chroma-js": "^2.4.2",
