+++ src/components/Tooltip/Tooltip.stories.tsx
@@ -1,47 +1,25 @@
 import { StoryObj } from '@storybook/react';
-import { Tooltip } from './Tooltip';
-import { TooltipProps } from './types';
-import { userEvent, within } from '@storybook/test';
-import { expect } from '@storybook/test';
-import { Button } from '..';
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from './Tooltip';
 
 export default {
   title: 'Components/Tooltip',
   component: Tooltip,
-  parameters: {
-    design: {
-      type: 'figma',
-      url: 'https://www.figma.com/file/dytYKVyXYobjZXq0BDPFsK/e2n-Web.21.09.23?node-id=1810%3A127553&mode=dev',
-    },
-  },
 };
 
-export const Default: StoryObj<TooltipProps> = {
-  render: (args: TooltipProps) => (
-    <Tooltip
-      {...args}
-      trigger={
-        <Button
-          type="button"
-          style={{ marginLeft: 150, marginTop: 50, marginBottom: 50 }}
-        >
-          Button
-        </Button>
-      }
-    >
-      Hier steht dein Tooltip Text!
-    </Tooltip>
+export const Default: StoryObj = {
+  render: () => (
+    <TooltipProvider>
+      <Tooltip>
+        <TooltipTrigger>Hover</TooltipTrigger>
+        <TooltipContent>
+          <p>Add to library</p>
+        </TooltipContent>
+      </Tooltip>
+    </TooltipProvider>
   ),
-  args: {
-    variant: 'default',
-    side: 'right',
-  },
-  play: async ({ canvasElement, step }) => {
-    const canvas = within(canvasElement);
-    await step('Tooltip container', async () => {
-      const tooltipComponent = canvas.getByTestId('tooltipContainer');
-      await expect(tooltipComponent).toBeInTheDocument();
-      await userEvent.hover(tooltipComponent);
-    });
-  },
 };
