+++ src/components/Tooltip/Tooltip.tsx
@@ -1,37 +1,28 @@
-import * as RadixUITooltip from '@radix-ui/react-tooltip';
-import { cx } from '@emotion/css';
-import { getStyles } from './styles';
-import { TooltipProps } from './types';
+import * as React from 'react';
+import * as TooltipPrimitive from '@radix-ui/react-tooltip';
 
-export function Tooltip({
-  variant = 'default',
-  side = 'bottom',
-  children,
-  trigger: Trigger,
-}: TooltipProps) {
-  const styles = getStyles(variant);
+import { cn } from '../../lib/utils';
 
-  return (
-    <div data-testid="tooltipContainer">
-      <RadixUITooltip.Provider>
-        <RadixUITooltip.Root>
-          <RadixUITooltip.Trigger asChild>{Trigger}</RadixUITooltip.Trigger>
-          <RadixUITooltip.Portal>
-            <RadixUITooltip.Content
-              side={side}
-              className={cx(styles.container, 'tooltip')}
-              sideOffset={2}
-            >
-              {children}
-              <RadixUITooltip.Arrow
-                width={'8px'}
-                height={'4px'}
-                className={cx(styles.arrow, 'arrow')}
-              />
-            </RadixUITooltip.Content>
-          </RadixUITooltip.Portal>
-        </RadixUITooltip.Root>
-      </RadixUITooltip.Provider>
-    </div>
-  );
-}
+const TooltipProvider = TooltipPrimitive.Provider;
+
+const Tooltip = TooltipPrimitive.Root;
+
+const TooltipTrigger = TooltipPrimitive.Trigger;
+
+const TooltipContent = React.forwardRef<
+  React.ElementRef<typeof TooltipPrimitive.Content>,
+  React.ComponentPropsWithoutRef<typeof TooltipPrimitive.Content>
+>(({ className, sideOffset = 4, ...props }, ref) => (
+  <TooltipPrimitive.Content
+    ref={ref}
+    sideOffset={sideOffset}
+    className={cn(
+      'z-50 overflow-hidden rounded-md border border-slate-200 bg-white px-3 py-1.5 text-sm text-slate-950 shadow-md animate-in fade-in-0 zoom-in-95 data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=closed]:zoom-out-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 dark:border-slate-800 dark:bg-slate-950 dark:text-slate-50',
+      className,
+    )}
+    {...props}
+  />
+));
+TooltipContent.displayName = TooltipPrimitive.Content.displayName;
+
+export { Tooltip, TooltipTrigger, TooltipContent, TooltipProvider };
