+++ src/components/Tooltip/index.ts
@@ -1,1 +1,1 @@
-export { Tooltip } from './Tooltip';
+export * from './Tooltip';
