+++ src/legacy/index.ts
@@ -6,3 +6,4 @@ export * from './Switch';
 export * from './RadioButton';
 export { LegacyTable as LegacyTable } from './Table';
 export * from './Tab';
+export * from './Tooltip';
